保理立项bug修改 ， 项目审批开发

+++ src/main/java/com/hand/hls/fct/dto/HlsCreditLineLease.java
@@ -47,7 +47,7 @@ public class HlsCreditLineLease extends BaseDTO {
      @Length(max = 32)
      private String leaseStatus;
 
-     private Float leasePrice;
+     private Double leasePrice;
 
      @Length(max = 512)
      private String description;
